allows plotting of points not just maps in tile_plot_proj, also fixed float to int conversion

+++ ecco_v4_py/resample_to_latlon.py
@@ -48,8 +48,8 @@ def resample_to_latlon(orig_lons, orig_lats, orig_field,
                                             lats=orig_lats_1d)
 
    # the latitudes to which we will we interpolate
-    num_lats = (new_grid_max_lat - new_grid_min_lat) / new_grid_delta_lat + 1
-    num_lons = (new_grid_max_lon - new_grid_min_lon) / new_grid_delta_lat + 1
+    num_lats = int((new_grid_max_lat - new_grid_min_lat) / new_grid_delta_lat) + 1
+    num_lons = int((new_grid_max_lon - new_grid_min_lon) / new_grid_delta_lat) + 1
 
     if (num_lats > 0) and (num_lons > 0):
         # linspace is preferred when using floats!
